+++ packages/register/register.ts
@@ -1,76 +1,86 @@
-import { platform } from 'os'
-import { resolve } from 'path'
-
-import { transformSync } from '@swc-node/core'
-import { SourcemapMap, installSourceMapSupport } from '@swc-node/sourcemap-support'
+import { Options, transform, transformSync } from '@swc-node/core'
+import { installSourceMapSupport, SourcemapMap } from '@swc-node/sourcemap-support'
 import { addHook } from 'pirates'
 import * as ts from 'typescript'
 
-import { readDefaultTsConfig } from './read-default-tsconfig'
+import { readDefaultTsConfig, tsCompilerOptionsToSwcConfig } from './read-default-tsconfig'
 
-const DEFAULT_EXTENSIONS = ['.js', '.jsx', '.es6', '.es', '.mjs', '.ts', '.tsx']
-const PLATFORM = platform()
+const DEFAULT_EXTENSIONS = new Set([
+  ts.Extension.Js,
+  ts.Extension.Ts,
+  ts.Extension.Jsx,
+  ts.Extension.Tsx,
+  ts.Extension.Mjs,
+  ts.Extension.Mts,
+  ts.Extension.Cjs,
+  ts.Extension.Cts,
+  '.es6',
+  '.es',
+])
 
-function toTsTarget(target: ts.ScriptTarget) {
-  switch (target) {
-    case ts.ScriptTarget.ES3:
-      return 'es3'
-    case ts.ScriptTarget.ES5:
-      return 'es5'
-    case ts.ScriptTarget.ES2015:
-      return 'es2015'
-    case ts.ScriptTarget.ES2016:
-      return 'es2016'
-    case ts.ScriptTarget.ES2017:
-      return 'es2017'
-    case ts.ScriptTarget.ES2018:
-      return 'es2018'
-    case ts.ScriptTarget.ES2019:
-    case ts.ScriptTarget.ES2020:
-    case ts.ScriptTarget.ESNext:
-    case ts.ScriptTarget.Latest:
-      return 'es2019'
-    case ts.ScriptTarget.JSON:
-      return 'es5'
+const injectInlineSourceMap = ({
+  filename,
+  code,
+  map,
+}: {
+  filename: string
+  code: string
+  map: string | undefined
+}): string => {
+  if (map) {
+    SourcemapMap.set(filename, map)
+    const base64Map = Buffer.from(map, 'utf8').toString('base64')
+    const sourceMapContent = `//# sourceMappingURL=data:application/json;charset=utf-8;base64,${base64Map}`
+    return `${code}\n${sourceMapContent}`
   }
+  return code
 }
 
-function toModule(moduleKind: ts.ModuleKind) {
-  switch (moduleKind) {
-    case ts.ModuleKind.CommonJS:
-      return 'commonjs'
-    case ts.ModuleKind.UMD:
-      return 'umd'
-    case ts.ModuleKind.AMD:
-      return 'amd'
-    case ts.ModuleKind.ES2015:
-    case ts.ModuleKind.ES2020:
-    case ts.ModuleKind.ESNext:
-    case ts.ModuleKind.None:
-      return 'es6'
-    case ts.ModuleKind.System:
-      throw new TypeError('Do not support system kind module')
-  }
-}
-
-function compile(
-  sourcecode: string,
+export function compile(
+  sourcecode: string | undefined,
   filename: string,
-  options: ts.CompilerOptions & { fallbackToTs?: (filename: string) => boolean },
+  options: ts.CompilerOptions & {
+    fallbackToTs?: (filename: string) => boolean
+  },
+): string
+
+export function compile(
+  sourcecode: string | undefined,
+  filename: string,
+  options: ts.CompilerOptions & {
+    fallbackToTs?: (filename: string) => boolean
+  },
+  async: false,
+): string
+
+export function compile(
+  sourcecode: string | undefined,
+  filename: string,
+  options: ts.CompilerOptions & {
+    fallbackToTs?: (filename: string) => boolean
+  },
+  async: true,
+): Promise<string>
+
+export function compile(
+  sourcecode: string | undefined,
+  filename: string,
+  options: ts.CompilerOptions & {
+    fallbackToTs?: (filename: string) => boolean
+  },
+  async: boolean,
+): string | Promise<string>
+
+export function compile(
+  sourcecode: string | undefined,
+  filename: string,
+  options: ts.CompilerOptions & {
+    fallbackToTs?: (filename: string) => boolean
+  },
+  async = false,
 ) {
-  if (filename.endsWith('.d.ts')) {
-    return ''
-  }
-  if (options.files && (options.files as string[]).length) {
-    if (
-      PLATFORM === 'win32' &&
-      (options.files as string[]).every((file) => filename !== resolve(process.cwd(), file))
-    ) {
-      return sourcecode
-    }
-    if (PLATFORM !== 'win32' && (options.files as string[]).every((file) => !filename.endsWith(file))) {
-      return sourcecode
-    }
+  if (sourcecode == null) {
+    return
   }
   if (options && typeof options.fallbackToTs === 'function' && options.fallbackToTs(filename)) {
     delete options.fallbackToTs
@@ -78,51 +88,39 @@ function compile(
       fileName: filename,
       compilerOptions: options,
     })
-    if (sourceMapText) {
-      SourcemapMap.set(filename, sourceMapText)
+    return injectInlineSourceMap({ filename, code: outputText, map: sourceMapText })
+  }
+
+  let swcRegisterConfig: Options
+  if (process.env.SWCRC) {
+    // when SWCRC environment variable is set to true it will use swcrc file
+    swcRegisterConfig = {
+      swc: {
+        swcrc: true,
+      },
     }
-    return outputText
   } else {
-    const { code, map } = transformSync(sourcecode, filename, {
-      target: toTsTarget(options.target ?? ts.ScriptTarget.ES2018),
-      module: toModule(options.module ?? ts.ModuleKind.ES2015),
-      sourcemap: createSourcemapOption(options),
-      jsx: filename.endsWith('.tsx') || filename.endsWith('.jsx') || Boolean(options.jsx),
-      react:
-        options.jsxFactory || options.jsxFragmentFactory
-          ? {
-              pragma: options.jsxFactory,
-              pragmaFrag: options.jsxFragmentFactory,
-            }
-          : undefined,
-      experimentalDecorators: options.experimentalDecorators ?? false,
-      emitDecoratorMetadata: options.emitDecoratorMetadata ?? false,
-      dynamicImport: options.module ? options.module >= ts.ModuleKind.ES2020 : true,
-      esModuleInterop: options.esModuleInterop ?? false,
-      keepClassNames: true,
+    swcRegisterConfig = tsCompilerOptionsToSwcConfig(options, filename)
+  }
+
+  if (async) {
+    return transform(sourcecode, filename, swcRegisterConfig).then(({ code, map }) => {
+      return injectInlineSourceMap({ filename, code, map })
     })
-    // in case of map is undefined
-    if (map) {
-      SourcemapMap.set(filename, map)
-    }
-    return code
+  } else {
+    const { code, map } = transformSync(sourcecode, filename, swcRegisterConfig)
+    return injectInlineSourceMap({ filename, code, map })
   }
 }
 
-export function register(options = readDefaultTsConfig()) {
+export function register(options: Partial<ts.CompilerOptions> = {}, hookOpts = {}) {
+  if (!process.env.SWCRC) {
+    options = Object.keys(options).length ? options : readDefaultTsConfig()
+  }
+  options.module = ts.ModuleKind.CommonJS
   installSourceMapSupport()
-  addHook((code, filename) => compile(code, filename, options), {
-    exts: DEFAULT_EXTENSIONS,
+  return addHook((code, filename) => compile(code, filename, options), {
+    exts: Array.from(DEFAULT_EXTENSIONS),
+    ...hookOpts,
   })
-}
-
-// @internal
-export function createSourcemapOption(options: ts.CompilerOptions) {
-  return options.sourceMap !== false
-    ? options.inlineSourceMap
-      ? 'both'
-      : true
-    : options.inlineSourceMap
-    ? 'inline'
-    : false
 }