+++ packages/jest/CHANGELOG.md
@@ -3,131 +3,335 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.2.1](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.2.0...@swc-node/jest@1.2.1) (2021-04-28)
+## [1.8.12](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.11...@swc-node/jest@1.8.12) (2024-07-17)
 
 **Note:** Version bump only for package @swc-node/jest
 
-# [1.2.0](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.1.1...@swc-node/jest@1.2.0) (2021-04-23)
+## [1.8.11](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.10...@swc-node/jest@1.8.11) (2024-07-16)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.10](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.9...@swc-node/jest@1.8.10) (2024-07-14)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.9](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.8...@swc-node/jest@1.8.9) (2024-07-13)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.8](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.7...@swc-node/jest@1.8.8) (2024-07-13)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.7](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.6...@swc-node/jest@1.8.7) (2024-07-12)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.6](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.5...@swc-node/jest@1.8.6) (2024-07-08)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.5](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.4...@swc-node/jest@1.8.5) (2024-07-05)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.4](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.3...@swc-node/jest@1.8.4) (2024-07-05)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.2...@swc-node/jest@1.8.3) (2024-06-28)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.1...@swc-node/jest@1.8.2) (2024-06-08)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.8.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.8.0...@swc-node/jest@1.8.1) (2024-05-01)
+
+### Bug Fixes
+
+- support compile js files. close [#761](https://github.com/swc-project/swc-node/issues/761) ([#767](https://github.com/swc-project/swc-node/issues/767)) ([016f1aa](https://github.com/swc-project/swc-node/commit/016f1aab2a17d2512d30b5a12848ed1941b59e49))
+
+# [1.8.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.9...@swc-node/jest@1.8.0) (2024-03-05)
 
 ### Features
 
-- **jest:** support jest 27 transformerConfig ([e2a51b2](https://github.com/Brooooooklyn/node-swc/commit/e2a51b2873898dcfc0a1b03a49f48ce70ea70f0a))
+- upgrade dependencies ([#751](https://github.com/swc-project/swc-node/issues/751)) ([653bd13](https://github.com/swc-project/swc-node/commit/653bd13c4ac84bd4bd28b886dc0f4e77362d0734))
 
-## [1.1.1](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.1.0...@swc-node/jest@1.1.1) (2021-03-04)
-
-**Note:** Version bump only for package @swc-node/jest
-
-# [1.1.0](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.0.3...@swc-node/jest@1.1.0) (2021-01-25)
+# [1.7.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.9...@swc-node/jest@1.7.0) (2024-02-01)
 
 ### Features
 
-- **jest:** cache transpile result to avoid source code to be compiled mult times ([77cc3d8](https://github.com/Brooooooklyn/node-swc/commit/77cc3d8ea82728b9b486e1b5fd898f82180c3f37))
+- upgrade dependencies ([#751](https://github.com/swc-project/swc-node/issues/751)) ([653bd13](https://github.com/swc-project/swc-node/commit/653bd13c4ac84bd4bd28b886dc0f4e77362d0734))
 
-## [1.0.3](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.0.2...@swc-node/jest@1.0.3) (2021-01-04)
-
-**Note:** Version bump only for package @swc-node/jest
-
-## [1.0.2](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.0.1...@swc-node/jest@1.0.2) (2020-11-18)
+## [1.6.9](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.8...@swc-node/jest@1.6.9) (2024-02-01)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [1.0.1](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@1.0.0...@swc-node/jest@1.0.1) (2020-10-21)
+## [1.6.8](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.7...@swc-node/jest@1.6.8) (2023-09-26)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.8](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.6...@swc-node/jest@0.3.8) (2020-09-16)
+## [1.6.7](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.6...@swc-node/jest@1.6.7) (2023-08-18)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.7](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.6...@swc-node/jest@0.3.7) (2020-09-13)
+## [1.6.6](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.5...@swc-node/jest@1.6.6) (2023-06-21)
+
+### Bug Fixes
+
+- tsCompilerOptionsToSwcConfig should not override default swc config for jest ([#714](https://github.com/swc-project/swc-node/issues/714)) ([60ea642](https://github.com/swc-project/swc-node/commit/60ea64284582ce3164ca3705976b4dc4215c2504))
+
+## [1.6.5](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.4...@swc-node/jest@1.6.5) (2023-04-24)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.6](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.5...@swc-node/jest@0.3.6) (2020-09-11)
+## [1.6.4](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.3...@swc-node/jest@1.6.4) (2023-04-10)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.5](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.4...@swc-node/jest@0.3.5) (2020-09-09)
+## [1.6.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.2...@swc-node/jest@1.6.3) (2023-03-31)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.4](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.3...@swc-node/jest@0.3.4) (2020-09-07)
+## [1.6.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.1...@swc-node/jest@1.6.2) (2023-02-15)
+
+### Bug Fixes
+
+- **core:** enforce sourcemap: 'inline' for jest ([#695](https://github.com/swc-project/swc-node/issues/695)) ([2439b7e](https://github.com/swc-project/swc-node/commit/2439b7e2cd28fee7d0d1038b91baeb042a32c146))
+
+## [1.6.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.6.0...@swc-node/jest@1.6.1) (2023-02-11)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.3.3](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.2...@swc-node/jest@0.3.3) (2020-09-06)
-
-**Note:** Version bump only for package @swc-node/jest
-
-## [0.3.2](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.1...@swc-node/jest@0.3.2) (2020-09-04)
-
-**Note:** Version bump only for package @swc-node/jest
-
-## [0.3.1](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.3.0...@swc-node/jest@0.3.1) (2020-09-01)
-
-**Note:** Version bump only for package @swc-node/jest
-
-# [0.3.0](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.27...@swc-node/jest@0.3.0) (2020-08-31)
+# [1.6.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.9...@swc-node/jest@1.6.0) (2023-02-10)
 
 ### Features
 
-- **core:** provide jest hoist plugin ([f3638f2](https://github.com/Brooooooklyn/node-swc/commit/f3638f2004b9fb323261a301b6fe354255846965))
-- **jest:** provide jest preset config ([7750ef5](https://github.com/Brooooooklyn/node-swc/commit/7750ef5d7cff3977a2b96f4d76810043e5f0988d))
+- **register:** experimental esm loader ([#643](https://github.com/swc-project/swc-node/issues/643)) ([0b4d305](https://github.com/swc-project/swc-node/commit/0b4d30505408f6f07c1ff8ea5c1953e1d22bb4e1))
 
-# [0.2.0](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.27...@swc-node/jest@0.2.0) (2020-08-28)
+## [1.5.9](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.8...@swc-node/jest@1.5.9) (2023-02-10)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.8](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.7...@swc-node/jest@1.5.8) (2023-02-07)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.7](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.6...@swc-node/jest@1.5.7) (2023-02-07)
+
+### Bug Fixes
+
+- **register:** always inline swc helpers ([1d557ec](https://github.com/swc-project/swc-node/commit/1d557ece0d9ccbba027ff9f2d262c03d4b918bcb))
+
+## [1.5.6](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.5...@swc-node/jest@1.5.6) (2023-01-05)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.5](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.4...@swc-node/jest@1.5.5) (2022-10-06)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.4](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.3...@swc-node/jest@1.5.4) (2022-10-06)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.2...@swc-node/jest@1.5.3) (2022-09-22)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [1.5.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.1...@swc-node/jest@1.5.2) (2022-05-05)
+
+### Performance Improvements
+
+- **jest:** readDefaultTsConfig should just execute once ([7223356](https://github.com/swc-project/swc-node/commit/7223356b820502627d4ab4a6e24a6ec698b9d031))
+
+## [1.5.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.5.0...@swc-node/jest@1.5.1) (2022-04-28)
+
+**Note:** Version bump only for package @swc-node/jest
+
+# [1.5.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.4.3...@swc-node/jest@1.5.0) (2022-04-27)
 
 ### Features
 
-- **core:** provide jest hoist plugin ([f3638f2](https://github.com/Brooooooklyn/node-swc/commit/f3638f2004b9fb323261a301b6fe354255846965))
+- **jest:** read tsconfig for default jest transform options ([8c180e6](https://github.com/swc-project/swc-node/commit/8c180e68abbc66aa68f83b401d985a6c8617baa9))
 
-## [0.1.27](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.26...@swc-node/jest@0.1.27) (2020-08-24)
-
-**Note:** Version bump only for package @swc-node/jest
-
-## [0.1.26](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.25...@swc-node/jest@0.1.26) (2020-08-23)
+## [1.4.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.4.2...@swc-node/jest@1.4.3) (2021-12-15)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.25](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.24...@swc-node/jest@0.1.25) (2020-08-23)
+## [1.4.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.4.1...@swc-node/jest@1.4.2) (2021-12-13)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.24](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.23...@swc-node/jest@0.1.24) (2020-08-21)
+## [1.4.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.4.0...@swc-node/jest@1.4.1) (2021-11-14)
+
+### Bug Fixes
+
+- **jest:** internal cache logic duplicated with jest cache ([43a36b9](https://github.com/swc-project/swc-node/commit/43a36b9be9aec0295a4a8d549fa189bef095d44f))
+
+# [1.4.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.5...@swc-node/jest@1.4.0) (2021-11-08)
+
+### Features
+
+- accept to transpile files with .mjs extension ([c1df0c8](https://github.com/swc-project/swc-node/commit/c1df0c827b8f8abaa7c5b04d837884378438385a))
+
+## [1.3.5](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.4...@swc-node/jest@1.3.5) (2021-10-29)
+
+### Bug Fixes
+
+- **jest:** xxhash should be dependency ([5559136](https://github.com/swc-project/swc-node/commit/555913638c07d17470d80a9944cba9e7a18bd94c))
+
+## [1.3.4](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.3...@swc-node/jest@1.3.4) (2021-10-29)
+
+### Performance Improvements
+
+- **jest:** replace sha1 with xxhash for better performance ([8d436dd](https://github.com/swc-project/swc-node/commit/8d436dd7000b9fee21c278bd294c017216e02b69))
+
+## [1.3.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.2...@swc-node/jest@1.3.3) (2021-10-16)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.23](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.22...@swc-node/jest@0.1.23) (2020-08-20)
+## [1.3.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.1...@swc-node/jest@1.3.2) (2021-09-11)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.22](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.21...@swc-node/jest@0.1.22) (2020-08-20)
+## [1.3.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.3.0...@swc-node/jest@1.3.1) (2021-06-07)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.20](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.19...@swc-node/jest@0.1.20) (2020-08-18)
+# [1.3.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.2.1...@swc-node/jest@1.3.0) (2021-05-21)
+
+### Features
+
+- **loader:** implement tsconfig compatible loader ([8c1cd85](https://github.com/swc-project/swc-node/commit/8c1cd858a64a6b6ec6ff23811bafab7dfe30554d))
+
+## [1.2.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.2.0...@swc-node/jest@1.2.1) (2021-04-28)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.19](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.18...@swc-node/jest@0.1.19) (2020-08-14)
+# [1.2.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.1.1...@swc-node/jest@1.2.0) (2021-04-23)
+
+### Features
+
+- **jest:** support jest 27 transformerConfig ([e2a51b2](https://github.com/swc-project/swc-node/commit/e2a51b2873898dcfc0a1b03a49f48ce70ea70f0a))
+
+## [1.1.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.1.0...@swc-node/jest@1.1.1) (2021-03-04)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.18](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.17...@swc-node/jest@0.1.18) (2020-08-13)
+# [1.1.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.0.3...@swc-node/jest@1.1.0) (2021-01-25)
+
+### Features
+
+- **jest:** cache transpile result to avoid source code to be compiled mult times ([77cc3d8](https://github.com/swc-project/swc-node/commit/77cc3d8ea82728b9b486e1b5fd898f82180c3f37))
+
+## [1.0.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.0.2...@swc-node/jest@1.0.3) (2021-01-04)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.17](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.16...@swc-node/jest@0.1.17) (2020-08-10)
+## [1.0.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.0.1...@swc-node/jest@1.0.2) (2020-11-18)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.16](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.15...@swc-node/jest@0.1.16) (2020-08-10)
+## [1.0.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@1.0.0...@swc-node/jest@1.0.1) (2020-10-21)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.15](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.14...@swc-node/jest@0.1.15) (2020-08-09)
+## [0.3.8](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.6...@swc-node/jest@0.3.8) (2020-09-16)
 
 **Note:** Version bump only for package @swc-node/jest
 
-## [0.1.14](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/jest@0.1.13...@swc-node/jest@0.1.14) (2020-08-09)
+## [0.3.7](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.6...@swc-node/jest@0.3.7) (2020-09-13)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.6](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.5...@swc-node/jest@0.3.6) (2020-09-11)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.5](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.4...@swc-node/jest@0.3.5) (2020-09-09)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.4](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.3...@swc-node/jest@0.3.4) (2020-09-07)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.3](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.2...@swc-node/jest@0.3.3) (2020-09-06)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.2](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.1...@swc-node/jest@0.3.2) (2020-09-04)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.3.1](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.3.0...@swc-node/jest@0.3.1) (2020-09-01)
+
+**Note:** Version bump only for package @swc-node/jest
+
+# [0.3.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.27...@swc-node/jest@0.3.0) (2020-08-31)
+
+### Features
+
+- **core:** provide jest hoist plugin ([f3638f2](https://github.com/swc-project/swc-node/commit/f3638f2004b9fb323261a301b6fe354255846965))
+- **jest:** provide jest preset config ([7750ef5](https://github.com/swc-project/swc-node/commit/7750ef5d7cff3977a2b96f4d76810043e5f0988d))
+
+# [0.2.0](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.27...@swc-node/jest@0.2.0) (2020-08-28)
+
+### Features
+
+- **core:** provide jest hoist plugin ([f3638f2](https://github.com/swc-project/swc-node/commit/f3638f2004b9fb323261a301b6fe354255846965))
+
+## [0.1.27](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.26...@swc-node/jest@0.1.27) (2020-08-24)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.26](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.25...@swc-node/jest@0.1.26) (2020-08-23)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.25](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.24...@swc-node/jest@0.1.25) (2020-08-23)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.24](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.23...@swc-node/jest@0.1.24) (2020-08-21)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.23](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.22...@swc-node/jest@0.1.23) (2020-08-20)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.22](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.21...@swc-node/jest@0.1.22) (2020-08-20)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.20](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.19...@swc-node/jest@0.1.20) (2020-08-18)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.19](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.18...@swc-node/jest@0.1.19) (2020-08-14)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.18](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.17...@swc-node/jest@0.1.18) (2020-08-13)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.17](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.16...@swc-node/jest@0.1.17) (2020-08-10)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.16](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.15...@swc-node/jest@0.1.16) (2020-08-10)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.15](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.14...@swc-node/jest@0.1.15) (2020-08-09)
+
+**Note:** Version bump only for package @swc-node/jest
+
+## [0.1.14](https://github.com/swc-project/swc-node/compare/@swc-node/jest@0.1.13...@swc-node/jest@0.1.14) (2020-08-09)
 
 **Note:** Version bump only for package @swc-node/jest