+++ packages/sourcemap-support/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.8](https://github.com/Brooooooklyn/node-swc/compare/@swc-node/sourcemap-support@0.1.7...@swc-node/sourcemap-support@0.1.8) (2020-08-31)
+## [0.5.1](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.5.0...@swc-node/sourcemap-support@0.5.1) (2024-07-05)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+# [0.5.0](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.3.0...@swc-node/sourcemap-support@0.5.0) (2024-03-05)
+
+### Features
+
+- upgrade dependencies ([#751](https://github.com/swc-project/swc-node/issues/751)) ([653bd13](https://github.com/swc-project/swc-node/commit/653bd13c4ac84bd4bd28b886dc0f4e77362d0734))
+
+# [0.4.0](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.3.0...@swc-node/sourcemap-support@0.4.0) (2024-02-01)
+
+### Features
+
+- upgrade dependencies ([#751](https://github.com/swc-project/swc-node/issues/751)) ([653bd13](https://github.com/swc-project/swc-node/commit/653bd13c4ac84bd4bd28b886dc0f4e77362d0734))
+
+# [0.3.0](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.2.4...@swc-node/sourcemap-support@0.3.0) (2023-02-10)
+
+### Features
+
+- **register:** experimental esm loader ([#643](https://github.com/swc-project/swc-node/issues/643)) ([0b4d305](https://github.com/swc-project/swc-node/commit/0b4d30505408f6f07c1ff8ea5c1953e1d22bb4e1))
+
+## [0.2.4](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.2.3...@swc-node/sourcemap-support@0.2.4) (2023-02-07)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+## [0.2.3](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.2.2...@swc-node/sourcemap-support@0.2.3) (2023-01-05)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+## [0.2.2](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.2.1...@swc-node/sourcemap-support@0.2.2) (2022-10-06)
+
+### Bug Fixes
+
+- **sourcemap-support:** add "tslib" as a dependnecy for "sourcemap-suport" ([#670](https://github.com/swc-project/swc-node/issues/670)) ([ec9ffeb](https://github.com/swc-project/swc-node/commit/ec9ffeb8691ac6c8d6b08c85f0f6890fc99004e9))
+
+## [0.2.1](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.2.0...@swc-node/sourcemap-support@0.2.1) (2022-09-22)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+# [0.2.0](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.1.11...@swc-node/sourcemap-support@0.2.0) (2022-04-27)
+
+### Features
+
+- **jest:** read tsconfig for default jest transform options ([8c180e6](https://github.com/swc-project/swc-node/commit/8c180e68abbc66aa68f83b401d985a6c8617baa9))
+
+## [0.1.11](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.1.10...@swc-node/sourcemap-support@0.1.11) (2021-12-13)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+## [0.1.10](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.1.9...@swc-node/sourcemap-support@0.1.10) (2021-10-29)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+## [0.1.9](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.1.8...@swc-node/sourcemap-support@0.1.9) (2021-10-16)
+
+**Note:** Version bump only for package @swc-node/sourcemap-support
+
+## [0.1.8](https://github.com/swc-project/swc-node/compare/@swc-node/sourcemap-support@0.1.7...@swc-node/sourcemap-support@0.1.8) (2020-08-31)
 
 **Note:** Version bump only for package @swc-node/sourcemap-support